Added dynamic resizing to background gradient
For some reason resizing the viewport caused the animation to not scale properly with the screen leading to users seeing the edges of the animation when they shouldn't have been able to.
It was my understanding that using percentages to define the properties of the background would've meant that it would update automatically to fit the screen size. Unfortunately this was not the case, as such I have added a small bit of JavaScript that forces the size to change whenever there is a viewport size change detected.